Time: 2 ms (38.76%), Space: 8.5 MB (53.69%) - LeetHub

+++ 0007-reverse-integer/0007-reverse-integer.cpp
@@ -14,6 +14,6 @@ public:
             return 0;
         }
 
-        return (int)rev;
+        return rev;
     }
 };
